server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv==1.0.1
openai
anthropic
langchain-anthropic
langchain-ollama
beautifulsoup4
requests
pypdf
firecrawl-py==0.0.20
